Updated RedirectView with UrlNotSetException in openwebpos application
In the RedirectView of openwebpos application, an exception UrlNotSetException has been introduced to handle the instances when a URL is not set. Previously, a TODO comment was left to handle this potential issue. Now instead of simply passing, this update includes raising of UrlNotSetException when the URL is not set.