Add templates folder creation to blueprint generation
This commit adds a new step to blueprint generation in the openwebpos cli tool. Now, apart from creating a forms directory and import in __init__.py, the new blueprint creation script also creates a templates directory for each new blueprint. This was necessary as each blueprint requires its own templates for proper functioning. This prevents missing template errors when a new blueprint is generated and used.