Add modals and buttons macros to UI templates
Introduced two new files modals.html and buttons.html to the openwebpos UI/templates/macros directory. These macros were added to simplify the creation and manipulation of modals and buttons throughout the application. It enables the team to maintain consistent design styles and streamline the process of making UI changes.