Correct application config settings in openwebpos
Updated the `config.py` in the openwebpos application to replace environment variable settings with static values. Removed the use of environment variables for debug, secret key, timezone, app path, etc. and replaced them with hard-coded values for consistency and to reduce dependency on the environment. Converted database configuration from dynamic to static; SQLite is now used in a consistent way. This may limit the flexibility of the application, but it increases predictability and strengthens secure handling of data.