"Added user Blueprint in Flask application
Introduced a new Blueprint instance for 'user' in the __init__.py file under the openwebpos/blueprints/user directory. This change streamlines the process of auto-registration of blueprints in the Flask application. It removes the need for manual additions in cli.py script, thus reducing possible errors and improving development efficiency."