Add SQLite datasource in project settings
The dataSources.xml file is added in the .idea directory to include the SQLite datasource called "openwebpos.db". This new datasource configuration is added to facilitate the integration of the SQLite database with the project. It provides the necessary settings such as the driver reference, synchronization status,and JDBC URI. This will allow easier and more efficient handling of SQL processes within the project.