Refactor buttons in the UI template
Refactored the macros in the 'buttons.html' file, defaulting the 'btn' macro to grey and 'icon' to list if no text is provided. Merged 'list_btn' into 'btn' for better code reusability and maintainability. Separating unique button types can create redundant codes that may be harder to manage. An update in the 'modal_btn' style is also introduced, sticking to a more consistent, smaller button throughout the application for a better user interface.