Add IngredientCategory view, form, model and template
This commits allows to user add, edit and remove Ingredient Categories in the recipes section of the application to provide better organization and management of the ingredients. New files for the ingredient category list view, form, models, and templates were created, and the ingredient category list view was registered in the product blueprint. The IngredientCategory model includes a unique name field and an active status.