Add ProductIngredient model and relationships
Created a new model, ProductIngredient, to build a many-to-many relationship between Product and Ingredient models. Modified Product and Ingredient models to include this relationship. Updated the __init__ file in product/models to import this new model. This change enables tracking of each product's ingredients, which will allow for more detailed product management and ingredient usage tracking in the POS system.