Add new form field macro in template
Introduced a new macro, select_field, in the form_fields.html template of the UI. This allows a more flexible use of form fields throughout the application, particularly helpful for an improved user interface in the OpenWebPos system. The new macro includes the field and its label within a div class, giving it a defined style.