Refactor URL for ProductIngredientDetailView route
The URL route for the ProductIngredientDetailView was refactored to simplify the URL structure, enhancing readability and making route navigation more intuitive. Previously, the URL was "product/<id>/ingredient", but it has been updated to "<id>/ingredient". This makes the URL structure more consistent with the other routes in the application and improves its maintainability.