Add field value option to select_field macro in form_fields
The 'select_field' macro in the 'form_fields.html' file has been updated to include an optional 'field_value' parameter. This allows for more dynamic rendering of form fields where the field's value is not known at the time of form definition but is only determined at run-time. If no 'field_value' is provided, the field functions as before, maintaining backwards compatibility.