Add ingredient amount and unit in product ingredient detail view
Ingredients associated with specific products can now contain detailed information including amount and unit. Changes have been made to the product_ingredient_detail.html template to display these additional details, and necessary fields have been added to the form in the ProductIngredientDetailView.py file.
The functionality to post and delete data within the form has also been updated accordingly. This enhancement will provide users with more detailed information about the ingredients used, including the quantity and measurement unit. This can be useful in scenarios where precise ingredient information is needed, such as in a restaurant or shop setting. Additionally, it also increases the flexibility and extensibility of the application, allowing for easier integration of new attributes or functionalities in the future.