Update ingredient choices formatting in ProductIngredientForm
Modified how ingredient choices are displayed in the ProductIngredientForm. Previously, the choices field was just populated with the ingredient id and name. However, with this update, each choice will now also display the ingredient's category name ahead of the ingredient name. This change will make it easier for users to identify and choose ingredients, especially when dealing with large lists.