Remove important font-size in td selector
The "!important" property was removed from the "font-size" attribute in the "td" selector in the corresponding CSS and SCSS files. This was done to improve the flexibility of the styling and prevent possible conflicts with other stylesheets that may be imported into the application. Allowing font-size modifications to the table data (td) elements can enhance readability in different screen sizes.