Add flexibility to queries in DetailView.py
Modified the DetailView.py to provide flexibility in the kind of queries conducted. Added two new attributes `url_variable` and `query_field` which allow specifying the field to query and the reference from the URL. Also updated `get_default_context` method to set navigation and page title based on the queried object's name. This change enhances the modularity and increases the utility of DetailView.