Add ingredient categories functionality
This commit introduces a new module within the application named 'ingredient'. The module includes model definitions for IngredientCategory and IngredientCategoryType, the blueprint and routing for '/ingredient/', and associated list and detail views. Templates for rendering the ingredient categories list and detail ID to the client browser also form part of this commit. This addition was necessitated by the need to better manage and categorize ingredients that are part of the products in the POS system.