Add ingredient templates to OpenWebPOS module
In this commit, a new path has been added to the OpenWebPOS module to manage the templates for 'ingredient'. This path will help in handling all the ingredient related categories templates under the module 'ingredient'. This change was necessary to categorize ingredients related to various products in the POS system.