Change dispatch_request parameters in BaseView
Updated the dispatch_request function in BaseView.py to pass only keyword arguments (kwargs). The previous version included args which was not necessary and could potentially cause errors if there were extra arguments passed to functions using dispatch_request. With this change, the function call is simplified and possible errors are minimized.