Add PUT method and enhance user experience in TemplateView
Implemented a PUT method in TemplateView to support update operations. This method retrieves the model by its id from the request form, then updates it with the cleaned form data. If the model hasn't been defined, it raises a NotImplementedError. Included the use of htmx_refresh method for seamless UI updates without requiring full page reloads. This consequently allows for improved interaction and responsiveness of the application.