Add get_by and get_all_by methods to sqlalchemy utility
Added two new methods to the sqlalchemy utility class in openwebpos. The 'get_by' method retrieves an instance of the class using the provided column and value, and the 'get_all_by' method retrieves all instances by the same parameters. These methods will allow for more flexible and efficient querying of the database.