Add URL variable in TemplateView class
The "url_variable" is added to the list of class attributes in the TemplateView.py of the openwebpos application. This new variable, set as "var", is used to retrieve an object from the URL. This change not only introduces a consistent approach to retrieve URL variables in the application but also lays the groundwork for further expansion of URL parsing features.