Refine query methods in DetailView class
Enhancements were made to the query methods in DetailView.py. The 'query_model' attribute was added to define the model to be queried. This change simplifies the query process as the model to be queried can now be explicitly defined inline, rather than having to query the entire model. If 'query_model' is not defined, an error is raised. Furthermore, commented out code has been removed to improve readability and maintainability.