Refactor class names in ingredient views
Updated class names of 'IngredientCategoryListView' and 'IngredientCategoryDetailView' to 'CategoryListView' and 'CategoryDetailView' respectively. Similarly, the corresponding imports and registers in '__init__.py' were updated accordingly. This change is done for code brevity and better readability. With only 'Category' in the name, it is still clear that these classes are about ingredient's category, not any other type of category.