Simplify class names in ingredient view
Refactored class names in the ingredient views to improve the code's readability and conciseness. The 'IngredientCategoryListView' and 'IngredientCategoryDetailView' are renamed to 'CategoryListView' and 'CategoryDetailView' respectively. This renaming still retains the intent of the classes, which are clearly about ingredient's category. Corresponding updates in '__init__.py' were also made according to these changes.